# crates/manta-cli/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manta API",
    "description": "REST API for managing CSM/OpenCHAMI HPC clusters via manta.",
    "contact": {
      "name": "Manuel Sopena Ballesteros",
      "email": "msopena@cscs.ch"
    },
    "license": {
      "name": ""
    },
    "version": "2.0.0-beta.64"
  },
  "servers": [
    {
      "url": "/api/v2",
      "description": "manta API v2"
    }
  ],
  "paths": {
...
    "/kernel-parameters": {
      "get": {
        "tags": [
          "kernel-parameters"
        ],
        "summary": "GET /kernel-parameters — fetch BSS kernel parameters for a group or node list.",
        "operationId": "get_kernel_parameters",
        "parameters": [
          {
            "name": "hsm_group",
            "in": "query",
            "description": "HSM group whose members' kernel parameters should be returned.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "nodes",
            "in": "query",
            "description": "Explicit comma-separated xnames; mutually exclusive with\n`hsm_group`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kernel parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BootParameters"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "kernel-parameters"
        ],
        "summary": "`DELETE /api/v2/kernel-parameters` — remove named kernel parameters from node BSS entries.",
        "operationId": "delete_kernel_parameters",
        "parameters": [
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteKernelParametersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Parameters removed or preview",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/kernel-parameters/add": {
      "post": {
        "tags": [
          "kernel-parameters"
        ],
        "summary": "`POST /api/v2/kernel-parameters/add` — merge new kernel parameters into existing node BSS entries.",
        "operationId": "add_kernel_parameters",
        "parameters": [
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddKernelParametersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Parameters added or preview",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/kernel-parameters/apply": {
      "post": {
        "tags": [
          "kernel-parameters"
        ],
        "summary": "`POST /api/v2/kernel-parameters/apply` — add, replace, or delete kernel parameters on nodes.",
        "operationId": "apply_kernel_parameters",
        "parameters": [
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyKernelParametersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Kernel parameters applied or preview",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddKernelParametersRequest": {
        "type": "object",
        "description": "Request body for `POST /api/v2/kernel-parameters/add` (append mode).",
        "required": [
          "params"
        ],
        "properties": {
          "dry_run": {
            "type": "boolean",
            "description": "When true, return the computed changeset without applying it."
          },
          "hsm_group": {
            "type": [
              "string",
              "null"
            ],
            "description": "Target HSM group; all members are resolved to xnames."
          },
          "overwrite": {
            "type": "boolean",
            "description": "When true, overwrite parameters that already exist."
          },
          "params": {
            "type": "string",
            "description": "Space-separated kernel parameter `key=value` pairs to add."
          },
          "project_sbps": {
            "type": "boolean",
            "description": "Whether to project SBPS images (default `true`)."
          },
          "xnames_expression": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hosts expression (xnames, NIDs, or hostlist notation); mutually\nexclusive with `hsm_group`."
          }
        }
      },
...
      "ApplyKernelParametersRequest": {
        "type": "object",
        "description": "Request body for `POST /api/v2/kernel-parameters/apply`.\n\nOne of `xnames_expression` or `hsm_group` must be set (not both).\nThe chosen [`KernelParamOp`] determines what `params` means:\nmerge-keys, replace-set, or remove-set.\n\n# Wire shape\n\n```json\n{\n  \"xnames_expression\": \"x3000c0s1b0n[0-3]\",\n  \"hsm_group\": null,\n  \"operation\": \"add\",\n  \"params\": \"console=ttyS0 nosmt\",\n  \"overwrite\": false,\n  \"project_sbps\": true,\n  \"dry_run\": false\n}\n```",
        "required": [
          "operation",
          "params"
        ],
        "properties": {
          "dry_run": {
            "type": "boolean",
            "description": "When true, return the computed changeset without applying it."
          },
          "hsm_group": {
            "type": [
              "string",
              "null"
            ],
            "description": "Target HSM group; all members are resolved to xnames."
          },
          "operation": {
            "$ref": "#/components/schemas/KernelParamOp",
            "description": "Which mutation to perform: add, apply (replace), or delete."
          },
          "overwrite": {
            "type": "boolean",
            "description": "Only relevant for the `Add` operation."
          },
          "params": {
            "type": "string",
            "description": "Space-separated kernel parameter `key=value` pairs."
          },
          "project_sbps": {
            "type": "boolean",
            "description": "Whether to project SBPS images (default `true`)."
          },
          "xnames_expression": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hosts expression (xnames, NIDs, or hostlist notation); mutually\nexclusive with `hsm_group`."
          }
        }
      },
...
      "BootParameters": {
        "type": "object",
        "properties": {
          "cloud-init": {},
          "hosts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "initrd": {
            "type": "string"
          },
          "kernel": {
            "type": "string"
          },
          "macs": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "nids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          "params": {
            "type": "string"
          }
        }
      },
...
      "DeleteKernelParametersRequest": {
        "type": "object",
        "description": "Request body for `DELETE /api/v2/kernel-parameters`.",
        "required": [
          "params"
        ],
        "properties": {
          "dry_run": {
            "type": "boolean",
            "description": "When true, return the computed changeset without applying it."
          },
          "hsm_group": {
            "type": [
              "string",
              "null"
            ],
            "description": "Target HSM group; all members are resolved to xnames."
          },
          "params": {
            "type": "string",
            "description": "Space-separated parameter names (or `key=value` pairs) to\nremove."
          },
          "xnames_expression": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hosts expression (xnames, NIDs, or hostlist notation); mutually\nexclusive with `hsm_group`."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard JSON error body returned by all failed endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable explanation of the failure. Never includes\nstack traces, credentials, or internal type names."
          }
        }
      },
...
      "KernelParamOp": {
        "type": "string",
        "description": "Which kernel-parameter mutation to perform on\n`POST /api/v2/kernel-parameters/apply`.",
        "enum": [
          "add",
          "apply",
          "delete"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}